requirements.txt
tweepy>=4.14.0
google-generativeai>=0.8.0
apscheduler>=3.10.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
